mock approval and finish the whole process bingo~

+++ src/main/java/com/loan/approve/api/dto/UserCreditInfoDTO.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -20,7 +21,7 @@ public class UserCreditInfoDTO {
     private boolean hasCreditQuota;
     private Date validDateBegin;
     private Date validDateEnd;
-    private double quotaAmount;
-    private double usedAmount;
+    private BigDecimal quotaAmount;
+    private BigDecimal usedAmount;
     private String currency;
 }
