+++ src/main/java/com/loan/approve/model/UserCreditAccount.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -18,8 +19,8 @@ public class UserCreditAccount {
     private boolean hasCreditQuota;
     private Date validDateBegin;
     private Date validDateEnd;
-    private double quotaAmount;
-    private double usedAmount;
+    private BigDecimal quotaAmount;
+    private BigDecimal usedAmount;
     private String currency;
 
 }
